pyproject.toml
[tool.poetry]
name = "babbagecoin"
version = "0.1.0"
description = "Second edition of project Babbage, create a blockchain from scratch (in a weekend)"
authors = ["Quentin Garchery <garchery.quentin@gmail.com>", "Martial Garchery <martial.garchery@gmail.com>", "Yohann Le Faou <yohannlefa@gmail.com>", "Julien Raspaud <julien.raspaud@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.3"
requests = "^2.27.1"
marshmallow = "^3.14.1"
cryptography = "^36.0.1"
python-dotenv = "^0.19.2"
sentry-sdk = {extras = ["flask"], version = "^1.5.8"}
redis = "^4.3.0"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
flake8 = "^4.0.1"
pytest = "^7.1.1"
pytest-cov = "^3.0.0"

[tool.black]
line-length = 100

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
